feat: session 제거 로직 추가, login 페이지 url 이동
