+++ src/pages/index.tsx
@@ -1,137 +1,68 @@
-import Head from 'next/head';
-import Image from 'next/image';
-import { Inter } from 'next/font/google';
-import styles from '@/styles/Home.module.css';
-import { useSession, signIn, signOut } from 'next-auth/react';
+import { IconButton, Box, Container as MContainer } from '@mui/material';
+import styled from '@emotion/styled';
+import GoogleLogo from '@/assets/google_login.svg';
+import KakaoLogo from '@/assets/kakao_login.svg';
+import ClaonLogo from '@/assets/claon_login_logo.svg';
+import { signIn } from 'next-auth/react';
+import { useOAuthSignIn } from '@/features/Login/queries';
+import { useEffect } from 'react';
 
-const inter = Inter({ subsets: ['latin'] });
+const LoginBox = styled(Box)`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: white;
+  width: 20rem;
+  height: 20rem;
+  padding: 1.25rem 0 3.125rem;
+  border-radius: 1.5rem;
+`;
 
-export default function Home() {
-  const { data: session } = useSession();
+const Text = styled.p`
+  font-size: 1.25rem;
+  color: black;
+  text-align: center;
+`;
+
+const Container = styled(MContainer)`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+`;
+
+export default function Login() {
+  const { signInClaon, session } = useOAuthSignIn();
+  useEffect(() => {
+    if (session) {
+      const { provider } = session;
+      signInClaon(provider);
+    }
+  }, [session]);
+
   return (
-    <>
-      {console.log(session)}
-      <button type="button" onClick={() => signIn('kakao')}>
-        KAKAO Sign in
-      </button>
-      <br />
-      <button type="button" onClick={() => signIn('google')}>
-        GOOGLE Sign in
-      </button>
-      <br />
-      <button type="button" onClick={() => signOut()}>
-        Sign out
-      </button>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className={styles.main}>
-        <div className={styles.description}>
-          <p>
-            Get started by editing&nbsp;
-            <code className={styles.code}>src/pages/index.tsx</code>
-          </p>
+    <main>
+      <Container maxWidth="sm" fixed sx={{ bgcolor: '#f2f2f2' }}>
+        <LoginBox>
+          <ClaonLogo />
           <div>
-            <a
-              href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              By{' '}
-              <Image
-                src="/vercel.svg"
-                alt="Vercel Logo"
-                className={styles.vercelLogo}
-                width={100}
-                height={24}
-                priority
-              />
-            </a>
+            <Text>
+              간편하게 로그인하고
+              <br />
+              클라온을 시작해 보세요!
+            </Text>
           </div>
-        </div>
-
-        <div className={styles.center}>
-          <Image
-            className={styles.logo}
-            src="/next.svg"
-            alt="Next.js Logo"
-            width={180}
-            height={37}
-            priority
-          />
-          <div className={styles.thirteen}>
-            <Image
-              src="/thirteen.svg"
-              alt="13"
-              width={40}
-              height={31}
-              priority
-            />
+          <div>
+            <IconButton onClick={() => signIn('google')}>
+              <GoogleLogo />
+            </IconButton>
+            <IconButton onClick={() => signIn('kakao')}>
+              <KakaoLogo />
+            </IconButton>
           </div>
-        </div>
-
-        <div className={styles.grid}>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Docs <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Find in-depth information about Next.js features and&nbsp;API.
-            </p>
-          </a>
-
-          <a
-            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Learn <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Learn about Next.js in an interactive course with&nbsp;quizzes!
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Templates <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Discover and deploy boilerplate example Next.js&nbsp;projects.
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Deploy <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Instantly deploy your Next.js site to a shareable URL
-              with&nbsp;Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-    </>
+        </LoginBox>
+      </Container>
+    </main>
   );
 }
